move dropdown utils to hook utils
