+++ src/hooks/useCombobox/utils.js
@@ -3,8 +3,12 @@ import PropTypes from 'prop-types'
 import {isControlledProp} from '../../utils.legacy'
 import {getState} from '../../utils'
 import {getInitialState as getInitialStateCommon} from '../utils.legacy'
-import {dropdownDefaultProps, dropdownPropTypes} from '../utils.dropdown'
-import {useIsInitialMount, useEnhancedReducer} from '../utils'
+import {
+  useIsInitialMount,
+  useEnhancedReducer,
+  dropdownDefaultProps,
+  dropdownPropTypes,
+} from '../utils'
 import {ControlledPropUpdatedSelectedItem} from './stateChangeTypes'
 
 export function getInitialState(props) {
