+++ src/hooks/useMultipleSelection/testUtils.js
@@ -1,7 +1,7 @@
 import * as React from 'react'
 
 import {render, screen, renderHook} from '@testing-library/react'
-import {dropdownDefaultProps} from '../utils.dropdown'
+import {dropdownDefaultProps} from '../utils'
 import {items, user, dataTestIds} from '../testUtils'
 import useCombobox from '../useCombobox'
 import {getInput, keyDownOnInput} from '../useCombobox/testUtils'
