+++ src/hooks/useSelect/reducer.js
@@ -1,12 +1,13 @@
 import {getHighlightedIndex, getNonDisabledIndex} from '../../utils'
-import {getDefaultValue} from '../utils'
-import commonReducer from '../reducer'
 import {
+  getDefaultValue,
   dropdownDefaultStateValues,
   getDefaultHighlightedIndex,
   getChangesOnSelection,
   getHighlightedIndexOnOpen,
-} from '../utils.dropdown'
+} from '../utils'
+import commonReducer from '../reducer'
+
 import {getItemIndexByCharacterKey} from './utils'
 import * as stateChangeTypes from './stateChangeTypes'
