--- src/hooks/utils.dropdown/index.ts
@@ -1,7 +1,0 @@
-export {propTypes as dropdownPropTypes} from './propTypes'
-export {defaultProps as dropdownDefaultProps} from './defaultProps'
-export {defaultStateValues as dropdownDefaultStateValues} from './defaultStateValues'
-export {useElementIds} from './useElementIds'
-export {getDefaultHighlightedIndex} from './getDefaultHighlightedIndex'
-export {getChangesOnSelection} from './getChangesOnSelection'
-export {getHighlightedIndexOnOpen} from './getHighlightedIndexOnOpen'
