+++ src/hooks/utils/__tests__/getDefaultHighlightedIndex.test.ts
