+++ src/hooks/utils/__tests__/getHighlightedIndexOnOpen.test.ts
