+++ src/hooks/utils/dropdownDefaultStateValues.ts
@@ -1,4 +1,4 @@
-export const defaultStateValues = {
+export const dropdownDefaultStateValues = {
   highlightedIndex: -1,
   isOpen: false,
   selectedItem: null as unknown,
