+++ src/hooks/utils/getChangesOnSelection.ts
