+++ src/hooks/utils/getDefaultHighlightedIndex.ts
