+++ src/hooks/utils/getHighlightedIndexOnOpen.ts
