+++ src/hooks/utils/index.ts
@@ -9,4 +9,11 @@ export {capitalizeString} from './capitalizeString'
 export {getDefaultValue} from './getDefaultValue'
 export {getInitialValue} from './getInitialValue'
 export {useA11yMessageStatus} from './useA11yMessageStatus'
+export {dropdownPropTypes} from './dropdownPropTypes'
+export {dropdownDefaultProps} from './dropdownDefaultProps'
+export {dropdownDefaultStateValues} from './dropdownDefaultStateValues'
+export {useElementIds} from '../utils/useElementIds'
+export {getDefaultHighlightedIndex} from './getDefaultHighlightedIndex'
+export {getChangesOnSelection} from './getChangesOnSelection'
+export {getHighlightedIndexOnOpen} from './getHighlightedIndexOnOpen'
 export type {StateReducer, Props, Action, Reducer} from './index.types'
